+++ frontend/src/App.tsx
@@ -22,29 +22,33 @@ function App() {
         title = "Home";
         metaDescription = "Welcome to the Home Page";
         break;
+        case "/newpatient":
+          title = "New Patient";
+          metaDescription = "Add details for a new patient.";
+          break;
       case "/addpatient":
         title = "Add Patient";
         metaDescription = "Add a new patient to the system.";
         break;
-      case "/newpatient":
-        title = "New Patient";
-        metaDescription = "Add details for a new patient.";
-        break;
+        case "/patientprofile":
+          title = "Patient Profile";
+          metaDescription = "View and manage patient profile details.";
+          break;
       case "/newdr":
         title = "New Doctor";
         metaDescription = "Add details for a new doctor.";
         break;
+      case "/doctorprofile":
+        title = "Doctor Profile";
+        metaDescription = "View and manage doctor profile details.";
+        break;
+      case "/rxitem":
+        title = "Rx Item";
+        metaDescription = "View and manage Rx item details.";
+      break;
       case "/newrx":
-        title = "New Prescription";
+        title = "New Rx";
         metaDescription = "Create a new prescription.";
-        break;
-      case "/newrx1":
-        title = "New Prescription 1";
-        metaDescription = "Create a new prescription (alternate version).";
-        break;
-      case "/patientprofile":
-        title = "Patient Profile";
-        metaDescription = "View and manage patient profile details.";
         break;
       case "/rxitemprofile":
         title = "Rx Item Profile";
